requirements.txt
customtkinter
numpy
pillow